all: fix test-sim-benchmark script
Fix the test-sim-benchmark script to use the new uplink.

Fixing it required to update the uplink to receive an access as a
parameter rather than a satellite address, API Key and an encryption
key.

Also add to errors returned by S3 client the output of the command
because it helps to figure it out what has produced the error. This is
very helpful when the invoked binaries has changed, due to updates, to
figure out what's the problem.

Change-Id: I8aaeb8720df96a9ae0ff0c84dc18123201857a2f